% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_render.R
\name{readd_pkg_title}
\alias{readd_pkg_title}
\title{Re-add package title to an rspkgdown index.html}
\usage{
readd_pkg_title(pkgname = basename(getwd()))
}
\arguments{
\item{pkgname}{A character string specifying the package name. Defaults to 
the name of the current working directory.}
}
\description{
This function re-adds the package title to the `docs/index.html` file of a 
rspkgdown-generated site. It modifies the HTML structure by inserting the 
package title as a header at the beginning of the main content section.
}
\details{
Some interaction between pkgdown and the rspkgdown template removes the
 package title from the `docs/index.html` file. This function restores it.
}
