% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_render.R
\name{rewrap_content}
\alias{rewrap_content}
\title{Re-Wrap rspkgdown main page content into `div#site-main`}
\usage{
rewrap_content(pkgname = basename(getwd()))
}
\arguments{
\item{pkgname}{A character string specifying the package name. Defaults
to the name of the current working directory.}
}
\description{
This function modifies HTML files in the `docs/` directory by wrapping
 the `body>nav.navbar` and `body>div.container` elements into a new
 `div#site-main` element, as is required by the rspkgdown template.
}
\details{
Even though the rspkgdown template already provides a `div#site-main`,
 some pkgdown logic removes it. This function restores it.
}
