% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_link_subfolders}
\alias{tweak_link_subfolders}
\title{Tweak links in HTML files within subfolders.}
\usage{
tweak_link_subfolders(selector, folder = "docs")
}
\arguments{
\item{selector}{[`character(1)`] CSS selector used to identify elements whose
`src` or `href` attributes should be tweaked.}

\item{folder}{[`character(1)`] [optional] Path to the root folder containing
HTML files. Defaults to `"docs"`.}
}
\value{
[`NULL`] This function is called for its side effects and returns
 `NULL`.
}
\description{
This function patches issues where `pkgdown::tweak_link_absolute()` and the
`{{#site}}{{root}}{{/site}}` template do not work, typically when the user
has not set a URL in the pkgdown meta configuration. It locates all `.html`
files in subfolders of the specified folder and prepends `"../"` to the
`src` and `href` attributes of elements matching the given CSS selector.
Matching elements are found using `xml2::xml_find_all()`.
}
\details{
Be careful: this function restyles the html file.
}
