% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_navbar_id}
\alias{tweak_navbar_id}
\title{Add navbar id for accessibility in HTML files.}
\usage{
tweak_navbar_id(folder = "docs")
}
\arguments{
\item{folder}{[`character(1)`] [optional] Path to the root folder containing
HTML files. Defaults to `"docs"`.}
}
\value{
[`NULL`] This function is called for its side effects and returns
  `NULL`.
}
\description{
This function sets the `id` attribute of the navbar element in HTML files
within the specified folder to `"pkgdown-navbar"`. This is required for
the `aria` target of the show-navbar button to work correctly. The navbar
element is identified using an XPath query for elements with `id` equal to
`"rspkgdown-main"` and a `nav` child containing the `"navbar"` class.
Matching elements are found using `xml2::xml_find_all()`.
}
